Add printing summary format to config
