+++ tests/test_multioutput.py
@@ -14,7 +14,7 @@ from gpflow.kernels import RBF
 from gpflow.likelihoods import Gaussian
 from gpflow.models import SVGP
 from gpflow.config import default_jitter, default_float
-from gpflow.utilities.training import set_trainable
+from gpflow.utilities import set_trainable
 
 float_type = default_float()
 rng = np.random.RandomState(99201)
